Move files around to make a real package
