+++ src/crystalsystems/__init__.py
@@ -1,0 +1,4 @@
+try:
+    from ._version import __version__  # noqa: F401
+except ImportError:
+    __version__ = "unknown"
